Extract mic volume meter canvas into hook
Because, mocking global object(window) everywhere in detail quite
burdensome
