+++ src/App.test.jsx
@@ -11,6 +11,7 @@ import App from './App';
 jest.mock('react-redux');
 jest.mock('./services/speechRecognitionService.js');
 jest.mock('./services/instances/audioContext.instance.js');
+jest.mock('./hooks/volumeCanvas.js');
 
 global.HTMLMediaElement.prototype.pause = jest.fn();
 global.HTMLCanvasElement.prototype.getContext = jest.fn();
