+++ src/hooks/__mocks__/volumeCanvas.js
@@ -1,0 +1,4 @@
+export const useVolumeCanvas = () => ({ current: {} });
+
+// TODO: delete this
+export const xx = {};
